added get by index and get as list functions
added the ability to get a specific block at a certain index. this
function will determine if it should start from the beginning or the end
of the chain for its search making worst case lookup O(n/2). also added
the ability to return the chain as a normal array list.